Corrections descrived in https://github.com/apache/phoenix/pull/1342
