+++ phoenix-core/src/main/java/org/apache/phoenix/schema/types/PBinary.java
@@ -40,8 +40,8 @@ public class PBinary extends PBinaryBase {
             Integer actualScale, SortOrder actualModifier, Integer desiredMaxLength, Integer desiredScale,
             SortOrder expectedModifier) {
         PVarbinary.INSTANCE.coerceBytes(ptr, o, actualType, actualMaxLength, actualScale, actualModifier, desiredMaxLength, desiredScale, expectedModifier);
-        if (null != o && null != desiredMaxLength && null != expectedModifier ) {
-            pad(ptr, desiredMaxLength, expectedModifier);
+        if ( (  o != null || ptr.getLength()>0 ) && null != desiredMaxLength && null != expectedModifier ) {
+        pad(ptr, desiredMaxLength, expectedModifier);
         }
     }
